publishers/spx_gold_trend/pyproject.toml
[project]
name = "stock-signals-daily"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloud-sql-python-connector>=1.20.0",
    "db-dtypes>=1.4.4",
    "gcsfs>=2026.2.0",
    "google-cloud>=0.34.0",
    "google-cloud-bigquery>=3.40.0",
    "google-cloud-bigquery-storage>=2.36.0",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pg8000>=1.31.5",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.0",
]
